Separation of the two solutions into two files
